增加iot_ota_newcore，iot_ota_newapp，iot_os_mem_used接口；变更iot_vat_send_cmd_push->iot_vat_push_cmd

+++ components/openat_inc/am_openat_drv.h
@@ -1598,8 +1598,8 @@ typedef struct
     union
     {
         struct pwm_pwt{
-        	UINT8  level;
-        	UINT32 freq;
+        	UINT16  level;
+        	UINT16 freq;
         }pwt,pwl;/*注意:使用PWL实现蜂鸣器时，freq实际有效范围:0-0xff*/
         struct pwm_lpg{
         	E_OPENAT_PWM_LPG_PERIOD period;
